Fix memory leak

+++ bot/handlers.py
@@ -2,21 +2,19 @@ from aiogram import Bot, Dispatcher
 from aiogram.client.default import DefaultBotProperties
 from aiogram.enums import ParseMode
 from aiogram.filters import CommandStart
-from aiogram.types import Chat, Message, User
-
-from spotdl import Song
-from spotipy.exceptions import SpotifyException
+from aiogram.types import Chat, FSInputFile, Message, User
 
 from core.settings import BOT_TOKEN
 from spotify import spotify
 
 from .middlewares import CreateUserMiddleware
-from .utils import download_and_send_song
+from .session import ResilientSession
 
-import asyncio
+from pathlib import Path
 
 bot = Bot(
     token=BOT_TOKEN,
+    session=ResilientSession(),
     default=DefaultBotProperties(
         parse_mode=ParseMode.HTML, link_preview_is_disabled=True
     ),
@@ -44,21 +42,21 @@ async def start_command_handler(message: Message, event_from_user: User) -> None
 
 @dispatcher.message()
 async def message_handler(message: Message, event_chat: Chat) -> None:
-    assert message.text
+    if not message.text or not message.text.startswith(''):
+        return None
 
     bot_message: Message = await message.reply(
         'Downloading music from the link...\nThis process may take a few minutes.'
     )
 
-    try:
-        songs: list[Song] = await spotify.search([message.text])
-    except SpotifyException:
+    songs: list[Path] | None = await spotify.download(message.text)
+
+    if not songs:
         await bot_message.edit_text(
             'Could not find or download music at the link, please try again later or send a different link.'
         )
         return
 
-    await asyncio.gather(
-        *[download_and_send_song(bot, event_chat, message, song) for song in songs]
-    )
+    for song in songs:
+        await message.reply_audio(FSInputFile(song))
     await bot_message.delete()
